requirements.txt
flask
python-dotenv
psycopg2-binary
supabase
flask-cors